requirements.txt
matplotlib==3.4.3
numpy==1.21.3
pandas==1.4.1
Pillow==9.3.0
plotly==5.6.0
scikit_learn==1.1.3
seaborn==0.11.2
streamlit==1.8.0
